+++ controllers/controller.Aspirants.js
@@ -1,108 +1,34 @@
-const models = require('../models')
-const axios = require('axios').default
+const { aspirantsLib, fundPool } = require('../lib'),
+  { getAll, search } = aspirantsLib,
+  { registAspirant } = fundPool;
 
 const aspirantsController = {};
 
-//--------------------------------------------------------------------
-// ------------------------Aspirants Utilities------------------------
-//--------------------------------------------------------------------
-// get all aspirant
-aspirantsController.getAll = async () => {
-  const aspirants = await models.Aspirants.findAll({})
-  return aspirants
-}
-
-// get by id
-aspirantsController.search = async (aspirantId) => {
-  const aspirants = await models.Aspirants.findOne({
-    where: {
-      storeKeeperId: aspirantId
-    }
-  })
-  return aspirants
-}
-
-
-// Register new aspirant
-aspirantsController.create = async (aspirant_id, email) => {
-  try {
-    const aspirants = await models.Aspirants.create({
-      storeKeeperId: aspirant_id,
-      email: email
-    })
-    return aspirants
-  } catch (error) {
-
-    if (error.errors[0].type == 'unique violation') return aspirantsController.search(aspirant_id)
-    console.error(error);
-  }
-}
-
-/**
- * Verify if a given email is a courier
- * 
- * returns 'none' if it is not a courier, and 'courier' in otherwise.
- */
-aspirantsController.verifyEmail = async (email) => {
-  try {
-    const response = await axios.get(`http://microservices.dev.rappi.com/api/rt-auth-helper/user/type?email=${email}`);
-    return response.data.user_type;
-  } catch (error) {
-    console.log(error);
-  }
+aspirantsController.getAll = async (req, res, next) => {
+  const aspirants = await getAll()
+  return res.json({
+    status: 'ok',
+    data: aspirants
+  });
 };
 
-/**
- * getAspirantToken get the aspirant token
- * @param {*} email 
- * @param {*} password 
- * 
- * return undefine if it was not possible to get the token, otherwise, returns the token.
- */
-aspirantsController.getToken = async (email, password) => {
-  const data = {
-    'client_id': '74HzD01JbhZ44iE1kh7Gt6dfNjEKrtWiz0FqTUDQ',
-    'client_secret': 'W8dOKF1mdHaG9wBNyoOCEBgHajO66GEl81lTDu2P',
-    'username': email,
-    'password': password,
-    'scope': 'all'
-  }
-  const options = {
-    'method': 'POST',
-    'headers': {
-      'uuid': '550e8400-e29b-41d4-a716-4466554400001234567',
-      'platform': 2,
-      'Content-Type': 'application/json'
-    }
-  }
-
-  try {
-    const response = await axios.post('http://microservices.dev.rappi.com/api/login/storekeeper', data, options);
-    return response.data.access_token;
-  } catch (error) {
-    //console.log(error);
-  }
+aspirantsController.search = async (req, res, next) => {
+  const { aspirant, fund } = await search(req.params.aspirant_id)
+  console.log(aspirant)
+  const status = (aspirant) ? 'ok' : 'not_found';
+  return res.json({
+    status: status,
+    aspirant: aspirant,
+    fund: fund
+  });
 };
 
-
-/**
- * 
- * @param {*} token 
- */
-aspirantsController.getStoreKeeperId = async (token) => {
-  const options = {
-    headers: {
-      'Content-Type': 'application/json',
-      'Authorization': `Bearer ${token}`,
-    }
-  }
-  try {
-    const response = await axios.get('http://microservices.dev.rappi.com/api/storekeepers-ms/storekeeper/rappitendero/profile?cache=false', options)
-    return response.data.id;
-  } catch (error) {
-    //console.log(error);
-  }
-}
-
+aspirantsController.registAspirant = async (req, res, next) => {
+  data = req.body;
+  if (!data.email || !data.password) return res.status(400).json({ 'message': 'Bad body' });
+  let result = await registAspirant(data.email, data.password);
+  console.log(result)
+  return res.json(result);
+};
 
 module.exports = aspirantsController;
