+++ lib/lib.aspirants.js
@@ -1,4 +1,5 @@
 const { Aspirants } = require('../models'),
+  { search } = require('./lib.funds'),
   axios = require('axios').default;
 
 const aspirantLib = { name: 'aspirantsLib' }
@@ -10,12 +11,13 @@ aspirantLib.getAll = async () => {
 
 // get by id
 aspirantLib.search = async (aspirantId) => {
-  const aspirants = await Aspirants.findOne({
+  const aspirant = await Aspirants.findOne({
     where: {
       storeKeeperId: aspirantId
     }
-  })
-  return aspirants
+  });
+  const fund = await search(aspirantId);
+  return { aspirant, fund };
 }
