+++ lib/lib.funds.js
@@ -38,6 +38,7 @@ fundsLib.getGathering = async () => {
 fundsLib.search = async (aspirant_id) => {
   try {
     return await Funds.findOne({
+      attributes: ['id', 'amount', 'status', 'createdAt'],
       where: {
         aspirant_id: aspirant_id
       }
