+++ tests/config.js
@@ -1,1 +1,1 @@
-exports.BASE_URL = `http://localhost:3000/api/v1`;
+exports.BASE_URL = `http://localhost:3001/api/v1`;
